+++ tests/plugins/pack/macapp/src/AppDelegate.m
@@ -1,0 +1,28 @@
+//
+//  AppDelegate.m
+//  macapp
+//
+//  Created by ruki on 2020/4/4.
+//  Copyright © 2020 tboox. All rights reserved.
+//
+
+#import "AppDelegate.h"
+
+@interface AppDelegate ()
+
+@end
+
+@implementation AppDelegate
+
+- (void)applicationDidFinishLaunching:(NSNotification *)aNotification {
+    // Insert code here to initialize your application
+}
+
+
+- (void)applicationWillTerminate:(NSNotification *)aNotification {
+    // Insert code here to tear down your application
+}
+
+
+@end
+